moved cpu, memody and disk components in a new components folder.
