+++ src/app/components/cpu/cpu.component.html
