+++ src/app/components/cpu/cpu.component.spec.ts
