+++ src/app/components/cpu/cpu.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { CpuInfo } from '../datastructures/system-information.datastructure';
+import { CpuInfo } from '../../datastructures/system-information.datastructure';
 
 @Component({
   selector: 'app-cpu',
