+++ src/app/components/disk/disk.component.html
