+++ src/app/components/disk/disk.component.spec.ts
