+++ src/app/components/memory/memory.component.html
