+++ src/app/components/memory/memory.component.spec.ts
